+++ octodns_cloudflare/__init__.py
@@ -47,6 +47,11 @@ class CloudflareRateLimitError(CloudflareError):
         CloudflareError.__init__(self, data)
 
 
+class Cloudflare5xxError(CloudflareError):
+    def __init__(self, data):
+        CloudflareError.__init__(self, data)
+
+
 _PROXIABLE_RECORD_TYPES = {'A', 'AAAA', 'ALIAS', 'CNAME'}
 
 
@@ -91,6 +96,7 @@ class CloudflareProvider(BaseProvider):
         pagerules=True,
         retry_count=4,
         retry_period=300,
+        auth_error_retry_count=0,
         zones_per_page=50,
         records_per_page=100,
         min_ttl=120,
@@ -124,6 +130,7 @@ class CloudflareProvider(BaseProvider):
         self.pagerules = pagerules
         self.retry_count = retry_count
         self.retry_period = retry_period
+        self.auth_error_retry_count = auth_error_retry_count
         self.zones_per_page = zones_per_page
         self.records_per_page = records_per_page
         self.min_ttl = min_ttl
@@ -140,6 +147,7 @@ class CloudflareProvider(BaseProvider):
 
     def _try_request(self, *args, **kwargs):
         tries = self.retry_count
+        auth_tries = self.auth_error_retry_count
         while True:  # We'll raise to break after our tries expire
             try:
                 return self._request(*args, **kwargs)
@@ -152,6 +160,28 @@ class CloudflareProvider(BaseProvider):
                     'for %ds and trying again, %d remaining',
                     self.retry_period,
                     tries,
+                )
+                sleep(self.retry_period)
+            except CloudflareAuthenticationError:
+                if auth_tries <= 0:
+                    raise
+                auth_tries -= 1
+                self.log.warning(
+                    'authentication error encountered, pausing '
+                    'for %ds and trying again, %d remaining',
+                    self.retry_period,
+                    auth_tries,
+                )
+                sleep(self.retry_period)
+            except Cloudflare5xxError:
+                if tries <= 0:
+                    raise
+                tries -= 1
+                self.log.warning(
+                    'http 502 error encountered, pausing '
+                    'for %ds and trying again, %d remaining',
+                    self.retry_period,
+                    auth_tries,
                 )
                 sleep(self.retry_period)
 
@@ -170,7 +200,8 @@ class CloudflareProvider(BaseProvider):
             raise CloudflareAuthenticationError(resp.json())
         if resp.status_code == 429:
             raise CloudflareRateLimitError(resp.json())
-
+        if resp.status_code in [502, 503]:
+            raise Cloudflare5xxError("http 5xx")
         resp.raise_for_status()
         return resp.json()